pyproject.toml
[project]
name = "aircraft-engine-remaining-life-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.9.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-docx>=1.2.0",
    "seaborn>=0.13.2",
]
